Modify the conditions of the coordinate comparing
On some mobile phone, I usually get two different but close start and end point values(eventData.pageX and eventData.event.pageX) when I touch on the same point, like "27.5" and "26.9687654125". In this case, the coordinate comparing will be "false", so that the "tap" event won't be triggered.

I modified the conditions of the coordinates comparing. Now it will be "true" when the end point is within 7 pixels from the start point.